+++ client/src/App.tsx
@@ -14,6 +14,8 @@ import Error404 from "./pages/Error404";
 import MonacoCode from "./pages/MonacoCode";
 import Tutorials from "./pages/Tutorials";
 import Admin from "./pages/Admin";
+import AdminQuestion from "./pages/AdminQuestion";
+import AdminTestcase from "./pages/AdminTestcase";
 
 export default function App() {
 
@@ -22,7 +24,10 @@ export default function App() {
       <Router>
         <Routes>
           <Route path="/" element={<AlgoVisualOpening/>}/>
-          <Route path="/admin" element={<Admin/>}></Route>
+          <Route path="/admin" element={<Admin/>}/>
+          <Route path="/admin/question" element={<AdminQuestion/>}/>
+          <Route path="/admin/testcase" element={<AdminTestcase/>}/>
+          {/* <Route path="/admin/courses" element={<Admin/>}/> */}
           <Route path="/landing" element={<Landing/>}/>
           <Route path="/accounts" element={<Accounts/>}/>
           <Route path="/home" element={<Home/>}/>
